solution/requirements.txt
torch>=1.10
torchvision>=0.11
numpy
Pillow
opencv-python
tensorboard
